// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  MEMBER: 'MEMBER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Country = {
  INDIA: 'INDIA',
  USA: 'USA'
} as const

export type Country = (typeof Country)[keyof typeof Country]


export const OrderStatus = {
  CREATED: 'CREATED',
  PLACED: 'PLACED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]